Update tensors-and-shapes.ipynb
I might be wrong, but from my understanding this should be the correct answers for those two. Tensors with shapes (4,2) and (4,) should not be able to be broadcasted from my understanding, so the second sensor should have shape (2,). I don't see a way for the other two to be broadcasted.